Add support for int32 and bool in DNP3 master + fix in codegen.

+++ codegen/src/main/java/com/oes/openfmb/generation/proto/TypeClassification.java
@@ -205,7 +205,7 @@ public class TypeClassification {
     private static String getInt32(FieldPath path)
     {
         if(path.hasName("setVal", "ctlVal", "stVal", "value")) {
-            return Types.int64.mapped;
+            return Types.int32.mapped;
         }
 
         throw new NoClassificationException(path);
